rename: API 파일 폴더명 변경
